[SBL-124] 설문 결과 페이지를 설문 관리페이지의 통계 보기 탭으로 이동, 바뀐 결과 API에 맞게 업데이트
