+++ src/app/management/[id]/page.tsx
@@ -30,7 +30,7 @@ export default function Page({ params }: { params: { id: string } }) {
   };
 
   let content;
-  if (tab === 0) content = <Tab0 />;
+  if (tab === 0) content = <Tab0 surveyId={id} />;
   else if (tab === 1) content = <Tab1 />;
   else if (tab === 2) content = <Tab2 />;
   else if (tab === 3) content = <Tab3 />;
