+++ src/services/result/fetch.ts
@@ -1,10 +1,10 @@
 import { kyWrapper } from '../ky-wrapper';
 import { makeUrl } from '../utils';
-import type { Results, SurveysResultParams } from './types';
+import type { SurveyResult, SurveysResultParams } from './types';
 
 const getResults = async ({ surveyId, resultFilter }: SurveysResultParams) => {
   const URL = makeUrl(['surveys', 'result', surveyId]);
-  return kyWrapper.post<Results>(URL, { json: resultFilter });
+  return kyWrapper.post<SurveyResult>(URL, { json: resultFilter });
 };
 
 export { getResults };
